Separate Clock and Quote in components inside Doing component fix/doing-structure
